feat: error parsing

+++ FinTrack.Server/Models/SecurityTransaction.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace FinTrack.Server.Models
 {
@@ -15,9 +14,9 @@ namespace FinTrack.Server.Models
         public int Id { get; set; }
         public OrderType OrderType { get; set; }
         public uint Quantity { get; set; }
-        public float Price { get; set; }
-        public float Commission { get; set; }
-        public float ExchangeRate { get; set; }
+        public decimal Price { get; set; }
+        public decimal Commission { get; set; }
+        public decimal ExchangeRate { get; set; }
         [MaxLength(3)] public string Currency { get; set; } = null!;
         public DateTime Date { get; set; } = DateTime.UtcNow;
         public virtual Security Security { get; set; } = null!;
